+++ CHANGELOG.md
@@ -1,7 +1,9 @@
-## Version 0.0.8
+## Version 0.1.0
 
 * Fixed bug in 'Show setup options' that made it difficult to return changed region settings back to the default (but any other change was ok)
-* Improved consistency of formatting used to display numbers based on Locale
+* Improved consistency of formatting used to display numbers based on other Locales
+* Switched default to US Locale
+* Removed pre-release notification
 
 
 ## Version 0.0.7